elemininate more warnings

+++ src/osgEarth/CacheSeed.cpp
@@ -86,7 +86,7 @@ void CacheSeed::seed( Map* map )
 
 			if (opt.minLevel().isSet() && opt.minLevel().get() < src_min_level)
                 src_min_level = opt.minLevel().get();
-			if (opt.maxLevel().isSet() && opt.maxLevel().get() > src_max_level)
+			if (opt.maxLevel().isSet() && opt.maxLevel().get() > (int)src_max_level)
                 src_max_level = opt.maxLevel().get();
         }
     }
@@ -120,7 +120,7 @@ void CacheSeed::seed( Map* map )
 
 			if (opt.minLevel().isSet() && opt.minLevel().get() < src_min_level)
                 src_min_level = opt.minLevel().get();
-			if (opt.maxLevel().isSet() && opt.maxLevel().get() > src_max_level)
+			if (opt.maxLevel().isSet() && opt.maxLevel().get() > (int)src_max_level)
                 src_max_level = opt.maxLevel().get();
 		}
     }
